eShop_backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
Requests==2.31.0
django-cors-headers==4.2.0
